OutageWatch/requirements.txt
azure-functions
requests
sendgrid
